requirements.txt
Django==1.11.3
django-crispy-forms==1.6.1
django-taggit==0.22.1
olefile==0.44
Pillow==4.2.1
pytz==2017.2
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2